6 new projects handed in
background slider - nice minimalist style background

drag and drop very easy to do with drag and drop components

drawing app alternative instead of using canvas. can be fixable but not recommended at the time to be used for personal projects

ripple effect coudn't find out a better and nice looking ripple but that is what we got ff

themeclock very nice project to work on, some degrees and  math and a nice looking clock is done. Anyways probably in future updates can add number around it. Thats all for today